vm -> API tests: consolidate testdata in dedicated folder
