+++ packages/vm/tests/api/EIPs/eip-2565-modexp-gas-cost.spec.ts
@@ -4,7 +4,7 @@ import Common from '@ethereumjs/common'
 import VM from '../../../lib'
 
 // See https://github.com/holiman/go-ethereum/blob/2c99023b68c573ba24a5b01db13e000bd9b82417/core/vm/testdata/precompiles/modexp_eip2565.json
-const testData = require('./eip-2565-testdata.json')
+const testData = require('../testdata/eip-2565.json')
 
 tape('EIP-2565 ModExp gas cost tests', (t) => {
   t.test('Test return data, gas cost and execution status against testdata', async (st) => {
